Prev button can restart song & code changes
-If less than 1s of the song has passed previous song gets played, if not the song is restarted
-moved seeking useEffect and states from PlayingBar into SongControls
