+++ src/components/PlayingBar/SongControls.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useState } from "react";
+import { useEffect, useRef, useState } from "react";
 import CustomRangeBar from "./CustomRangeBar";
 import SongControlsButtons from "./SongControlsButtons";
 import { AudioSettingsProp } from "../../App";
@@ -8,30 +8,27 @@ import { AudioSettingsProp } from "../../App";
 interface SongControlsProp {
   isPlaying: boolean;
   setIsPlaying: React.Dispatch<React.SetStateAction<boolean>>;
-  playback: number;
-  setPlayback: React.Dispatch<React.SetStateAction<number>>;
-  isDragging: boolean;
-  setIsDragging: React.Dispatch<React.SetStateAction<boolean>>;
   audioSettings: AudioSettingsProp | null;
   songDuration: number;
   songTime: number;
   setQueueIndex: React.Dispatch<React.SetStateAction<number>>;
+  initSong: (number: number) => void;
 }
 
 export default function SongControls({
   isPlaying,
   setIsPlaying,
-  playback,
-  setPlayback,
-  isDragging,
-  setIsDragging,
   audioSettings,
   songDuration,
   songTime,
   setQueueIndex,
+  initSong,
 }: SongControlsProp) {
   const id = "playback";
   const [songTimeEl, setSongTimeEl] = useState<number | undefined>(undefined);
+  const [isSeeking, setIsSeeking] = useState(false);
+  const [playback, setPlayback] = useState(0);
+  const isSeekingRef = useRef(false);
 
   //update song time element
   useEffect(() => {
@@ -40,13 +37,27 @@ export default function SongControls({
   }, [playback]);
 
   useEffect(() => {
-    if (isDragging) return;
+    if (isSeeking) return;
     setSongTimeEl(songTime);
 
     //song time goes from 1 to duration (e.g 180 [seconds])
     //1 / duration is like saying the 1th step of the range bar
     setPlayback(songTime / Math.floor(songDuration));
   }, [songTime]);
+
+  //handle seeking
+  useEffect(() => {
+    if (isSeeking) {
+      isSeekingRef.current = true;
+    }
+
+    if (!isSeeking && isSeekingRef.current) {
+      const currentSeekTime = Math.floor(playback * songDuration);
+      initSong(currentSeekTime);
+
+      isSeekingRef.current = false;
+    }
+  }, [isSeeking]);
 
   //format seconds to 0:00
   function formatSeconds(seconds: number) {
@@ -66,6 +77,8 @@ export default function SongControls({
           isAudioContext={audioSettings?.audioCtx}
           isSource={audioSettings?.source}
           setQueueIndex={setQueueIndex}
+          songTime={songTime}
+          initSong={initSong}
         />
       </div>
       <div className="flex w-full text-xs whitespace-nowrap">
@@ -77,8 +90,8 @@ export default function SongControls({
           steps={songDuration !== undefined ? Math.floor(songDuration) : 0}
           progress={playback}
           setProgress={setPlayback}
-          isDragging={isDragging}
-          setIsDragging={setIsDragging}
+          isDragging={isSeeking}
+          setIsDragging={setIsSeeking}
         />
         <span>{songDuration !== 0 ? formatSeconds(songDuration) : "-:--"}</span>
       </div>
